+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Added
+- `metrics-http-json.rb`: Added `-H/--header` flag to pass custom HTTP headers.
 
 
 ## [6.0.1] - 2020-01-30